PB-1727: add --max-attempts option to e2e-tests.
The CodeBuild client defaults to making no more than 3 attempts for each call.
Some changes trigger many calls to CodeBuild which throttles us to the point of
three attempts not being sufficient. In practice we see errors like this:

	2025/05/26 12:02:04 Starting E2E tests on dev staging
	2025/05/26 12:02:04  with tests: tests.service_alti
	2025/05/26 12:02:05 E2E tests started with ID: e2e-tests-dev-pr:131d177d-0577-4f9e-a354-5c8d7dc3f9d6
	2025/05/26 12:02:05 https://eu-central-1.console.aws.amazon.com/codesuite/codebuild/974517877189/projects/e2e-tests-dev-pr/build/e2e-tests-dev-pr%3A131d177d-0577-4f9e-a354-5c8d7dc3f9d6?region=eu-central-1
	␛[31mfailed to get build status: operation error CodeBuild: BatchGetBuilds, exceeded maximum number of attempts, 3, https response error StatusCode: 400, RequestID: e6edd560-70b2-4f82-a631-dac1920f8b88, api error ThrottlingException: Rate exceeded: BatchGetBuilds␛[0m

With this change, we allow the caller of e2e-tests to set an arbitrary number of
attempts with the --max-attempts commandline option. The default is still 3.

I verified this works by reproducing the problem manually like this:

```
$ seq 1 100 | while read ; do ./e2e-tests get --test-id $some_build_id & done
```

Doing the same with "--max-attempts=10", the command completes successfully
although it takes longer.