Define a sensible "all" Makefile target.
The default target is `help`.
The `all` target is an alias for `help`.
There is no target to build all the binaries.

This change updates `all` to build all the programs as is [generally expected](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html#Standard-Targets).

This allows simplifying both developer workflow and the buildspec.